Feat/donating (#18)
* donation 낱개 조회

* feat/donating

* Donation 클래스의 targetAmount와 currentAmount를 int에서 Long 타입으로 변경함

* DonationController에 기부하기(donating) 기능 구현함

* tmp

* tmp

* MemberController에 test() 삭제

* HttpServletRequest

* Principal을 HttpServletRequest로 변경

+++ src/main/java/wintersprouts/boogie/domain/donation/ApplyDonationForm.java
@@ -18,7 +18,7 @@ public class ApplyDonationForm {
     private String content;
 
     @NotNull(message = "목표 금액은 필수 사항입니다. ")
-    private int targetAmount;
+    private Long targetAmount;
 
     @Future(message = "마감일은 과거일 수 없습니다. ")
     private LocalDate dueDate;
